requirements.txt
beautifulsoup4==4.12.3
PyPDF2==3.0.1
python-dotenv==1.0.1
haystack-ai==2.9.0
streamlit==1.41.1